# adt: split interval tree by right endpoint on matched left endpoints