@serathius 2025-04-22
Please include benchmark from https://github.com/etcd-io/etcd/issues/19769 in the PR.