@redwrasse 2025-04-23
> Please include benchmark from #19769 in the PR.

Will do, I'll investigate using benchstats.