@redwrasse 2025-05-05
> > Please include benchmark from #19769 in the PR.
> 
> Will do, I'll investigate using benchstats.

Will try to get benchstats added in the next few days.