@serathius 2025-05-13
Can you add the benchmarks to the PR so we can use them to evaluate performance in the future? We could also consider merging them first in separate PR.