@serathius 2025-05-13
If there is no performance change for `Insert` maybe let's drop the changes there from the PR. Reducing the scope will make the PR easier to review. Complicating `Insert` code without improving performance doesn't seem like a good tradeoff.