@redwrasse 2025-05-13
If I'm understanding , are you suggesting removing the code changes to `Insert` but keeping the ones in `Find`? I'm not sure how that would work- the interval tree structure changes need to be reflected in both methods I believe.

I can definitely add in the benchmarks from the description, if we feel they are appropriate.