@redwrasse 2025-05-16
Cormen "Introduction to Algorithms", Chapter 14 Exercise 14.3.5 is I think relevant:

14.3-5
Suggest modifications to the interval-tree procedures to support the new operation `INTERVAL-SEARCH-EXACTLY(T,i)`, where `T` is an interval tree and `i` is an interval. The operation should return a pointer to a node `x` in `T` such that `x.int.low = i.low` and `x.int.high = i.high`, or `T.nil` if `T` contains no such node. All operations, including `INTERVAL-SEARCH-EXACTLY`, should run in `O(lg n)` time on an `n`-node interval tree.