@serathius 2025-05-17
Coverage looks pretty good https://app.codecov.io/gh/etcd-io/etcd/pull/19768?src=pr&el=tree&utm_medium=referral&utm_source=github&utm_content=comment&utm_campaign=pr+comments&utm_term=etcd-io