@ahrtr 2025-05-21
We should add a performance item in 3.7 changelog, let me take care of together with another change.